requirements.txt
fastapi
pygithub